Report Linux errnos from epoll_wait, mm, futex, shm and the path helpers
Fifth pass over the syscall surface, against /usr/src/linux 7.0-rc7. Every
change names the kernel function that decides it, in the code.

The remaining `-1` sentinels reach libc as EPERM, and this wave clears the
last dense pockets: epoll_wait, the memory-management family, futex/futex2,
SysV shm, capset/capget, groups, gethostname, and the readlink/link/mknod
helpers in core.inc.rs.

But the more interesting finds were not wrong errnos. They were arguments
accepted and then IGNORED, which is the silent half of the same problem:

  mmap            MAP_FIXED_NOREPLACE was ignored entirely, so a caller
                  asking "map here or fail" had its existing mapping
                  REPLACED. `do_mmap` gives -EEXIST. That is data loss, not
                  a wrong number.
  futex_waitv     entries were validated lazily, interleaved with reading
                  the futex words, so some words were read before a later
                  entry was rejected. `futex_parse_waitv` validates the whole
                  vector first.
  futex_requeue   the source entry's `val` is a COMPARE operand —
                  `cmpval = futexes[0].w.val` — and a moved word is -EAGAIN.
                  NARF skipped the compare and requeued against a stale view,
                  which is exactly the race a condvar uses it to detect.
  epoll_pwait     a wrong `sigsetsize` silently skipped the mask instead of
                  -EINVAL, so the signal the caller was blocking could be
                  delivered inside the wait.
  shmget          SHM_HUGETLB was accepted and ignored, handing back a
                  4 KiB-backed segment where `newseg` returns -EINVAL.
  mknod           S_IFDIR created a directory — a NARF-only extension no
                  portable caller can use. `may_mknod` returns -EPERM, and
                  screens the type BEFORE the path lookup.
  mprotect        undefined `prot` bits were accepted; several handlers read
                  `int` arguments from the whole 64-bit register.

epoll_wait's checks were also in the wrong ORDER: `do_epoll_wait` resolves
the descriptor first (-EBADF), then maxevents (-EINVAL), then access_ok
(-EFAULT), then is_file_epoll (-EINVAL). NARF tested the buffer before the
fd, so an event loop whose epfd had closed saw EPERM rather than the EBADF
that tells it to rebuild.

Three musl-demo smokes were themselves pinning NARF away from Linux, and are
corrected here rather than worked around: futex2 masks must fit the futex
width (0xffffffff, not ~0UL); futex_requeue with a stale compare value is
-EAGAIN; and `capget(&bad_version, NULL)` SUCCEEDS — it is the version probe
libcap uses before allocating, and reporting -EINVAL made it give up instead
of retrying at the reported version. The futex2 smoke now pins both halves
of the requeue compare so it cannot regress to a no-op.

Not every sentinel was wrong. `shmctl`'s four are genuinely -EPERM
(`ipcctl_obtain_check` is seeded `int err = -EPERM;`), and so are
`process_vm_readv`'s cross-AS denial (Linux explicitly maps EACCES to EPERM
there) and `mknod(S_IFDIR)`. Those were verified and left alone.

Four tests were pinning the old behaviour and are updated with the reason
recorded. One of them was passing by accident: `smoke_userspace_futex_wake_op_rmw`
used 0xF00D as a wait-queue key, which `get_futex_key`'s
`(address % size) != 0` check correctly rejects.

Local gates: syscall_abi + userspace 1946 pass / 0 fail; fmt clean; clippy
clean on x86_64 and aarch64 under -D warnings. The full suite has not been
run yet.

+++ userspace/src/handlers/sys_mlockall.rs
@@ -3,11 +3,33 @@ use super::*;
 
 /// `mlockall(flags)` — atomically replace the address space's future-lock
 /// policy and, when MCL_CURRENT is present, lock every existing ordinary VMA.
+///
+/// `mm/mlock.c::SYSCALL_DEFINE1(mlockall)`:
+///
+/// ```text
+///     SYSCALL_DEFINE1(mlockall, int, flags)
+///     if (!flags || (flags & ~(MCL_CURRENT | MCL_FUTURE | MCL_ONFAULT)) ||
+///         flags == MCL_ONFAULT)
+///             return -EINVAL;
+///     if (!can_do_mlock())
+///             return -EPERM;
+///     ... /* over RLIMIT_MEMLOCK with MCL_CURRENT and no CAP_IPC_LOCK => -ENOMEM */
+/// ```
+///
+/// `flags` is declared **`int`**, so the kernel only ever inspects the low 32
+/// bits; the upper half of the register is not part of the request. Reading
+/// the whole 64-bit argument made `mlockall(MCL_CURRENT)` fail with EINVAL
+/// whenever a caller left garbage in the high half of `rdi` — which a libc
+/// wrapper that stores the argument as `int` and sign-/zero-extends
+/// inconsistently, or a language runtime marshalling through a `long`, can
+/// easily do. EINVAL here reads as "this kernel does not support these
+/// MCL_ flags" and makes a runtime disable memory locking wholesale, where
+/// EPERM would have told it to ask for privilege instead.
 pub(crate) fn sys_mlockall(ctx: &mut dyn TrapContext) {
-    const MCL_CURRENT: u64 = 1;
-    const MCL_FUTURE: u64 = 2;
-    const MCL_ONFAULT: u64 = 4;
-    let flags = ctx.args().arg0;
+    const MCL_CURRENT: u32 = 1;
+    const MCL_FUTURE: u32 = 2;
+    const MCL_ONFAULT: u32 = 4;
+    let flags = ctx.args().arg0 as u32;
     if flags == 0
         || flags & !(MCL_CURRENT | MCL_FUTURE | MCL_ONFAULT) != 0
         || flags & MCL_ONFAULT != 0 && flags & (MCL_CURRENT | MCL_FUTURE) == 0
